# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseConfig": {
        "additionalProperties": false,
        "properties": {
          "agreement_rules": {
            "$ref": "#/components/schemas/AgreementRules"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "opponent_private_context": {
            "title": "Opponent Private Context",
            "type": "string"
          },
          "opponent_private_phrases": {
            "items": {
              "type": "string"
            },
            "title": "Opponent Private Phrases",
            "type": "array"
          },
          "opponent_role": {
            "title": "Opponent Role",
            "type": "string"
          },
          "opponent_strategy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpponentStrategy"
              },
              {
                "type": "null"
              }
            ]
          },
          "player_private_context": {
            "title": "Player Private Context",
            "type": "string"
          },
          "player_role": {
            "title": "Player Role",
            "type": "string"
          },
          "shared_context": {
            "title": "Shared Context",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "shared_context",
          "player_role",
          "opponent_role",
          "player_private_context",
          "opponent_private_context"
        ],
        "title": "CaseConfig",
        "type": "object"
      },
...
      "FinishRequest": {
        "additionalProperties": false,
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseConfig"
          },
          "preparation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreparationCard"
              },
              {
                "type": "null"
              }
            ]
          },
          "snapshot": {
            "$ref": "#/components/schemas/SessionSnapshot"
          }
        },
        "required": [
          "case",
          "snapshot"
        ],
        "title": "FinishRequest",
        "type": "object"
      },
      "FinishResponse": {
        "additionalProperties": false,
        "properties": {
          "judge_verdicts": {
            "items": {
              "$ref": "#/components/schemas/JudgeSlot"
            },
            "title": "Judge Verdicts",
            "type": "array"
          },
          "outcome": {
            "$ref": "#/components/schemas/OutcomeResult"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "trainer_feedback": {
            "$ref": "#/components/schemas/TrainerSlot"
          }
        },
        "required": [
          "session_id",
          "outcome",
          "judge_verdicts",
          "trainer_feedback"
        ],
        "title": "FinishResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JudgeCollege": {
        "enum": [
          "hiring",
          "negotiation",
          "ownership"
        ],
        "type": "string"
      },
      "JudgeSlot": {
        "additionalProperties": false,
        "properties": {
          "college": {
            "$ref": "#/components/schemas/JudgeCollege"
          },
          "error_code": {
            "anyOf": [
              {
                "enum": [
                  "judge_unavailable",
                  "invalid_judge_output"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "status": {
            "enum": [
              "ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JudgeVerdict"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "college",
          "status"
        ],
        "title": "JudgeSlot",
        "type": "object"
      },
...
      "ModelErrorCode": {
        "enum": [
          "invalid_opponent_output",
          "opponent_unavailable",
          "guard_unavailable",
          "invalid_guard_output",
          "guard_uncertain",
          "validator_unavailable",
          "invalid_validator_output",
          "validator_uncertain"
        ],
        "type": "string"
      },
      "NonEmptyText": {
        "minLength": 1,
        "type": "string"
      },
...
      "OpponentStrategy": {
        "additionalProperties": false,
        "properties": {
          "steps": {
            "items": {
              "$ref": "#/components/schemas/OpponentPositionStep"
            },
            "minItems": 3,
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "steps"
        ],
        "title": "OpponentStrategy",
        "type": "object"
      },
      "OutcomeResult": {
        "additionalProperties": false,
        "properties": {
          "agreement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DealTerms"
              },
              {
                "type": "null"
              }
            ]
          },
          "commitments": {
            "items": {
              "type": "string"
            },
            "title": "Commitments",
            "type": "array"
          },
          "kind": {
            "enum": [
              "agreement",
              "partial_agreement",
              "deferred",
              "no_agreement"
            ],
            "title": "Kind",
            "type": "string"
          },
          "next_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Step"
          },
          "open_points": {
            "items": {
              "type": "string"
            },
            "title": "Open Points",
            "type": "array"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "summary"
        ],
        "title": "OutcomeResult",
        "type": "object"
      },
...
      "PreparationCard": {
        "additionalProperties": false,
        "properties": {
          "arguments": {
            "items": {
              "$ref": "#/components/schemas/NonEmptyText"
            },
            "title": "Arguments",
            "type": "array"
          },
          "negotiation_goal": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negotiation Goal"
          },
          "planned_questions": {
            "items": {
              "$ref": "#/components/schemas/NonEmptyText"
            },
            "title": "Planned Questions",
            "type": "array"
          },
          "possible_solutions": {
            "items": {
              "$ref": "#/components/schemas/NonEmptyText"
            },
            "title": "Possible Solutions",
            "type": "array"
          },
          "situation_analysis": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Situation Analysis"
          },
          "strategic_goal": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategic Goal"
          }
        },
        "title": "PreparationCard",
        "type": "object"
      },
...
      "SessionSnapshot": {
        "additionalProperties": false,
        "properties": {
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/SessionState"
          },
          "transcript": {
            "items": {
              "$ref": "#/components/schemas/TranscriptEntry"
            },
            "title": "Transcript",
            "type": "array"
          }
        },
        "required": [
          "session_id",
          "state",
          "transcript"
        ],
        "title": "SessionSnapshot",
        "type": "object"
      },
...
      "TrainerSlot": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "enum": [
                  "trainer_unavailable",
                  "invalid_trainer_output"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrainerFeedback"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "TrainerSlot",
        "type": "object"
      },
...
      "TurnRequest": {
        "additionalProperties": false,
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseConfig"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SessionSnapshot"
          },
          "turn_id": {
            "title": "Turn Id",
            "type": "string"
          },
          "user_text": {
            "minLength": 1,
            "title": "User Text",
            "type": "string"
          }
        },
        "required": [
          "case",
          "snapshot",
          "turn_id",
          "user_text"
        ],
        "title": "TurnRequest",
        "type": "object"
      },
      "TurnResponse": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "opponent_text": {
            "title": "Opponent Text",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SessionSnapshot"
          },
          "status": {
            "enum": [
              "accepted",
              "blocked",
              "model_error"
            ],
            "title": "Status",
            "type": "string"
          },
          "turn_id": {
            "title": "Turn Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "turn_id",
          "status",
          "opponent_text",
          "snapshot"
        ],
        "title": "TurnResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Arena AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/finish": {
      "post": {
        "operationId": "finish_duel_v1_finish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Finish Duel"
      }
    },
...
    "/v1/turn": {
      "post": {
        "operationId": "take_turn_v1_turn_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Take Turn"
      }
    }
  }
}